refactor(sysmon): move analysis progress hook to src/workspaces/sysmon/
